Fix build with Qt 6.9 / newer clang
- Raise QT_MOST to 6.9.99 (builds and runs fine against Qt 6.9.2).
- Qt 6.9 removed char16_t from QString::arg integral overloads and
  QChar's implicit enum conversion; add explicit casts in textutils,
  resistor, and paletteitembase.

Note for macOS/Xcode 26 SDK builds: configure with
  qmake6 ... 'QMAKE_CXXFLAGS += -include arm_acle.h'
to work around Qt 6.9's qyieldcpu.h relying on the __yield intrinsic.

+++ src/items/paletteitembase.cpp
@@ -653,7 +653,7 @@ QString PaletteItemBase::retrieveSvg(ViewLayer::ViewLayerID viewLayerID, QHash<Q
 
 	//DebugDialog::debug(QString("path: %1").arg(path));
 
-	QString svg = svgHash.value(path + xmlName + QString(QChar(m_viewLayerPlacement)), "");
+	QString svg = svgHash.value(path + xmlName + QString(QChar((int) m_viewLayerPlacement)), "");
 	if (!svg.isEmpty()) return svg;
 
 	SvgFileSplitter splitter;
@@ -684,7 +684,7 @@ QString PaletteItemBase::retrieveSvg(ViewLayer::ViewLayerID viewLayerID, QHash<Q
 		return "";
 	}
 	svg = splitter.elementString(xmlName);
-	svgHash.insert(path + xmlName + QString(QChar(m_viewLayerPlacement)), svg);
+	svgHash.insert(path + xmlName + QString(QChar((int) m_viewLayerPlacement)), svg);
 	return svg;
 }
